test-runner: Avoid encoding/decoding strings as much as possible
Different Python versions handle string encodings in different ways,
so to avoid compatibility issues, read and write the raw bytes.
The logger is the only part that cares about strings being valid in
a particular encoding, so that is the one place the bytes are decoded
into a string.  In this case, invalid characters are replaced with a
replacement character.  In the other cases, logged messages are passed
unaltered to the "stdout", "stderr", and "merged" log files.