Cross-platform acltype
The acltype property is currently hidden on FreeBSD and does not
reflect the NFSv4 style ZFS ACLs used on the platform.  This makes it
difficult to observe that a pool imported from FreeBSD on Linux has a
different type of ACL that is being ignored, and vice versa.

Add an nfsv4 acltype and expose the property on FreeBSD.

Make the default acltype nfsv4 on FreeBSD.

Setting acltype to an unhandled style is treated the same as setting
it to off.  The ACLs will not be removed, but they will be ignored.

Rename posixacl to posix and prefer off to noacl throughout.  The
Linux mount options remain named posixacl and noacl to avoid ambiguity
and breaking existing configurations.  To the latter end, also retain
posixacl and noacl values for acltype as aliases.

Signed-off-by: Ryan Moeller <ryan@iXsystems.com>