Disambiguate condvar API contract.
On Illumos callers of cv_timedwait and cv_timedwait_hires
can't distinguish between whether or not the cv was signalled
or the call timed out. Illumos handles this (for some definition
of handles) by calling cv_signal in the return path if we were
signalled but the return value indicates instead that we timed
out. This would make sense if it were possible to query the the
cv for its net signal disposition. However, this isn't possible
and, in spite of the fact that there are places in the code that
clearly take a different and incompatible path if a timeout value
is indicated, this distinction appears to be rather subtle to most
developers. This problem is further compounded by the fact that on
Linux, calling cv_signal in the return path wouldn't even do the
right thing unless there are other waiters.

Since it is possible for the caller to independently determine how
much time is remaining but it is not possible to query if the cv
was in fact signalled, prioritizing signalling over timeout seems
like a cleaner solution. In addition, judging from usage patterns
within the code itself, it is also less error prone.

Signed-off-by: Matt Macy <mmacy@FreeBSD.org>