Restore DOS R/O attribute semantics on write
ZFS_READONLY should not be checked in zfs_write.  Instead it should be
treated like an ACL and not checked after the file has been opened R/W.

Port the portion of zfs_zaccess_common that enforces the attribute
semantics from FreeBSD to Linux and restore the behavior from FreeBSD
in zfs_write.

Signed-off-by: Ryan Moeller <ryan@iXsystems.com>